+++ src/ui/views/About.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 
-import LiveSplitIcon from "../assets/icon.svg";
-import { Markdown } from "../util/Markdown";
-import * as variables from "../css/variables.icss.scss";
+import LiveSplitIcon from "../../assets/icon.svg";
+import { Markdown } from "../../util/Markdown";
 import { ArrowLeft } from "lucide-react";
 
-import "../css/About.scss";
+import * as variables from "../../css/variables.icss.scss";
+import "../../css/About.scss";
 
 export interface Props {
     callbacks: Callbacks;
